requirements.txt
pycrypto==2.6.1
flask==0.12
requests==2.18.4